+++ formulario-tipo/inc/header.php
@@ -1,8 +1,8 @@
 <header class="u-borderBottom u-bgWhite u-py2">
 	<div class="Container">
-		<h1 class="u-h5 u-mb0 u-inlineBlock  u-pl2 u-md-pl0">
+		<span class="u-h5 u-mb0 u-inlineBlock  u-pl2 u-md-pl0">
 			<a href="https://www.gub.uy/agencia-gobierno-electronico-sociedad-informacion-conocimiento/componentes-reutilizables" class="u-textBlack">Componentes reutilizables</a>
-		</h1>
+</span>
 		
 		<div class="Header-parent u-inlineBlock">
 			<a class="u-textMute" href="https://www.gub.uy/agencia-gobierno-electronico-sociedad-informacion-conocimiento/">AGESIC</a>