+++ formulario-tipo/index.php
@@ -12,13 +12,21 @@
 
         <div class="Grid">
 
-            <div class="Grid-item Grid-item--center u-mt5 u-md-size3of4 u-lg-size8of10">
+            <div class="Grid-item Grid-item--center u-mt0 u-md-size3of4 u-lg-size8of10">
       
-								<div class="Page Page-document Page--formulario">
+								<div class="Page Page-document Page--formulario u-pt0">
 
-									<h2 class="Page-title">Formulario tipo</h2>
+									<h1 class="Page-title">Formulario tipo</h1>
+									
+									<p>El formulario tipo es un conjunto de elementos básicos, modelos de interacción y diseño visual para formularios, que puedan utilizarse como soluciones transversales a todos los organismos del Estado Uruguayo.</p>
+									<p>El Formulario tipo incluye:</p>
+									<ul>
+										<li>La <strong>estandarización del maquetado</strong> de los átomos (elementos indivisibles) que se utilizan para la creación de formularios en HTML.</li>
+										<li>La definición de un <strong>modelo único de interacción</strong> con los usuarios.</li>
+										<li>El <strong>diseño visual comsistente</strong> para la construcción de formularios en el Estado.</li>
+									</ul>
 
-									<h3>Elementos de formulario</h3>
+									<h2>Elementos de formulario</h2>
 
 									<div class="Box--listadoAccesos u-mt3">    
 										<ul class="Box-gridList Grid Grid--rowSize3">
@@ -30,10 +38,11 @@
 																	</div>
 																	<div class="Flag-body">
 																			<span class="Box-title">Navegación</span>
-																			<span class="Box-info"></span>
+																			<span class="Box-info">Pasos, botones y enlaces.</span>
 																	</div>
 															</div>
-													</a>
+												</a>
+													
 											</li>
 											<li class="Grid-item">
 													<a href="entrada-de-datos-inputs.php">
@@ -42,8 +51,8 @@
 																			<img src="./img/inputs.png" alt="">
 																	</div>
 																	<div class="Flag-body">
-																			<span class="Box-title">Entrada de datos (inputs)</span>
-																			<span class="Box-info"></span>
+																			<span class="Box-title">Entrada de datos</span>
+																			<span class="Box-info">Elementos interactivos (inputs).</span>
 																	</div>
 															</div>
 													</a>
@@ -56,7 +65,7 @@
 																	</div>
 																	<div class="Flag-body">
 																			<span class="Box-title">Presentación de datos y cuadros de diálogo</span>
-																			<span class="Box-info"></span>
+																			<span class="Box-info">Elementos informativos, mensajes y tablas.</span>
 																	</div>
 															</div>
 													</a>
@@ -65,7 +74,7 @@
 									</div>	
 											
 											
-									<h3>Ejemplos aplicados</h3>
+									<h2>Ejemplos aplicados</h2>
 
 									<div class="Box--listadoAccesos u-mt3">    
 										<ul class="Box-gridList Grid Grid--rowSize3">
@@ -77,7 +86,7 @@
 																	</div>
 																	<div class="Flag-body">
 																			<span class="Box-title">Ejemplo de formulario</span>
-																			<span class="Box-info"></span>
+																			<span class="Box-info">Con pasos, grupos, campos varios y acciones de ejemplo.</span>
 																	</div>
 															</div>
 													</a>
@@ -90,7 +99,7 @@
 																	</div>
 																	<div class="Flag-body">
 																			<span class="Box-title">Bloques reutilizables</span>
-																			<span class="Box-info"></span>
+																			<span class="Box-info">Grupos de campos estándar para solicitud de información.</span>
 																	</div>
 															</div>
 													</a>
@@ -104,7 +113,7 @@
 																	</div>
 																	<div class="Flag-body">
 																			<span class="Box-title">Trámite tipo</span>
-																			<span class="Box-info"></span>
+																			<span class="Box-info">Ejemplo de trámite en función de su contenido.</span>
 																	</div>
 															</div>
 													</a>
@@ -112,7 +121,7 @@
 										</ul>
 									</div>	
 											
-									<h3>Proyecto y ayudas</h3>
+									<h2>Proyecto y ayudas</h2>
 
 									<div class="Box--listadoAccesos u-mt3">    
 										<ul class="Box-gridList Grid Grid--rowSize3">
@@ -144,154 +153,6 @@
 											</li>
 										</ul>
 									</div>	
-
-
-									<!-- <div class="Box--listadoAccesos u-mt3">    
-										<ul class="Box-gridList Grid Grid--rowSize3">
-											<li class="Grid-item">
-													<a href="atomos.php">
-															<div class="Flag">
-																	<div class="Flag-image">
-																			<img src="./img/atomo.png" alt="">
-																	</div>
-																	<div class="Flag-body">
-																			<span class="Box-title">Atomos</span>
-																			<span class="Box-info"></span>
-																	</div>
-															</div>
-													</a>
-											</li>
-											<li class="Grid-item">
-													<a href="moleculas.php">
-															<div class="Flag">
-																	<div class="Flag-image">
-																			<img src="./img/moleculas.png" alt="">
-																	</div>
-																	<div class="Flag-body">
-																			<span class="Box-title">Moléculas</span>
-																			<span class="Box-info"></span>
-																	</div>
-															</div>
-													</a>
-											</li>
-											<li class="Grid-item">
-													<a href="bloques.php">
-															<div class="Flag">
-																	<div class="Flag-image">
-																			<img src="./img/bloques.png" alt="">
-																	</div>
-																	<div class="Flag-body">
-																			<span class="Box-title">Bloques</span>
-																			<span class="Box-info"></span>
-																	</div>
-															</div>
-													</a>
-											</li>
-											<li class="Grid-item">
-													<a href="formulario.php">
-															<div class="Flag">
-																	<div class="Flag-image">
-																			<img src="./img/ejemplo-completo.png" alt="">
-																	</div>
-																	<div class="Flag-body">
-																			<span class="Box-title">Ejemplo completo</span>
-																			<span class="Box-info"></span>
-																	</div>
-															</div>
-													</a>
-											</li>
-											<li class="Grid-item">
-													<a href="tramite-tipo-01.php">
-															<div class="Flag">
-																	<div class="Flag-image">
-																			<img src="./img/tramite-tipo.png" alt="">
-																	</div>
-																	<div class="Flag-body">
-																			<span class="Box-title">Trámite tipo</span>
-																			<span class="Box-info"></span>
-																	</div>
-															</div>
-													</a>
-											</li>
-											<li class="Grid-item">
-													<a href="https://github.com/AGESIC-UY/componentes-reutilizables-formulario-tipo">
-															<div class="Flag">
-																	<div class="Flag-image">
-																			<img src="./img/github.png" alt="">
-																	</div>
-																	<div class="Flag-body">
-																			<span class="Box-title">Repositorio Github</span>
-																			<span class="Box-info"></span>
-																	</div>
-															</div>
-													</a>
-											</li>
-										</ul>
-									</div>		 -->
-
-									<h3>Qué es el Formulario Tipo</h3>
-									<p>El formulario tipo incluye un conjunto de elementos básicos, modelos de interacción y diseño visual para formularios, que puedan utilizarse como soluciones transversales a todos los organismos del Estado Uruguayo.</p>
-									<p>El Formulario tipo incluye:</p>
-									<ul>
-										<li>La <strong>estandarización del maquetado</strong> de los átomos (elementos indivisibles) que se utilizan para la creación de formularios en HTML.</li>
-										<li>La definición de un <strong>modelo único de interacción</strong> con los usuarios.</li>
-										<li>El <strong>diseño visual genérico</strong> que se recomienda para la construcción de formularios del Estado.</li>
-									</ul>
-
-									<!--<h2 class="u-h2 u-mt5">Componentes</h2>
-
-									<h3 class="u-h4 u-mt5">Átomos</h2>
-									<ul>
-										<li><a href="atomos.php">Listado de átomos</a></li>
-									</ul>
-
-									<h3 class="u-h4 u-mt5">Moléculas</h2>
-									<ul>
-										<li><a href="moleculas.php">Listado de moléculas</a></li>
-									</ul>
-
-									<h3 class="u-h4 u-mt5">Bloques</h2>
-									<ul>
-										<li>
-											<a href="bloques.php">Listado de bloques</a>
-											<ul>
-												<li><a href="bloques-acerca-del-tramite.php">Acerca del trámite</a></li>
-												<li><a href="bloques-confirmacion.php">Confirmación</a></li>
-												<li><a href="bloques-consentimiento-informado.php">Consentimiento informado</a></li>
-												<li><a href="bloques-datos-de-contacto.php">Datos de contacto</a></li>
-												<li><a href="bloques-datos-de-domicilio.php">Datos de domicilio</a></li>
-												<li><a href="bloques-datos-de-domicilio-completo.php">Datos de domicilio completo</a></li>
-												<li><a href="bloques-firma.php">Métodos de firma</a></li>
-												<li><a href="bloques-identificacion-documento-uruguayo.php">Identificación con documento de identidad uruguayo</a></li>
-												<li><a href="bloques-identificacion-documento-extranjero.php">Identificación con documento de identidad extranjero</a></li>
-												<li><a href="bloques-ingreso-usuarios.php">Ingreso de usuario</a></li>
-												<li><a href="bloques-medios-de-pago.php">Medios de pago</a></li>
-												<li><a href="bloques-medios-de-pago-detalle.php">Medios de pago (detalle)</a></li>
-												<li><a href="bloques-pasos-pendientes.php">Pasos pendientes</a></li>
-												<li><a href="bloques-personas-juridicas.php">Representación de Personas jurídicas</a></li>
-												<li><a href="bloques-registro-notificaciones.php">Registro ante notificaciones.gub.uy</a></li>
-												<li><a href="bloques-valoracion.php">Valoración</a></li>
-												<li><a href="bloques-bandeja-de-entrada.php">Bandeja de entrada</a></li>
-											</ul>
-										</li>
-									</ul>
-
-									<h3>Ejemplo completo</h3>
-									<ul>
-										<li><a href="formulario.php">Atómos, moléculas y bloques aplicados</a></li>
-									</ul>
-
-									<h3>Trámite tipo</h3>
-									<ul>
-										<li><a href="tramite-tipo-01.php">Paso 1 - Datos personales</a></li>
-										<li><a href="tramite-tipo-02.php">Paso 2 - Datos del trámite</a></li>
-										<li><a href="tramite-tipo-03.php">Paso 3 - Firma</a></li>
-										<li><a href="tramite-tipo-04.php">Paso 4 - Agenda</a></li>
-										<li><a href="tramite-tipo-05.php">Paso 5 - Medios de pago</a></li>
-										<li><a href="tramite-tipo-06.php">Paso 6 - Confirmación</a></li>
-									</ul>
-									-->
-
 								</div>
 
             </div>